+++ packages/kas/src/__tests__/checking-form_test.js
@@ -1,11 +1,12 @@
+// @flow
 import _ from "underscore";
 
 import * as KAS from "../index.js";
 
 expect.extend({
     toHaveNorm(input: string, reference: string) {
-        var actual = KAS.parse(input).expr.normalize().print();
-        var expected = KAS.parse(reference).expr.normalize().print();
+        const actual = KAS.parse(input).expr.normalize().print();
+        const expected = KAS.parse(reference).expr.normalize().print();
 
         return {
             pass: actual === expected,
@@ -13,8 +14,8 @@ expect.extend({
         };
     },
     toHaveStripNorm(input: string, reference: string) {
-        var actual = KAS.parse(input).expr.strip().normalize().print();
-        var expected = KAS.parse(reference).expr.strip().normalize().print();
+        const actual = KAS.parse(input).expr.strip().normalize().print();
+        const expected = KAS.parse(reference).expr.strip().normalize().print();
 
         return {
             pass: actual === expected,
@@ -28,7 +29,7 @@ describe("checking form", () => {
         expect("ab").toHaveNorm("ba");
         expect("(ab)c").toHaveNorm("(cb)a");
 
-        var forms = [
+        const forms = [
             "(6x+1)(x-1)",
             "(1+6x)(x-1)",
             "(6x+1)(-1+x)",
@@ -36,10 +37,10 @@ describe("checking form", () => {
             "(x-1)(6x+1)",
             "(x-1)(1+6x)",
             "(-1+x)(6x+1)",
-            "(-1+x)(1+6x)"
+            "(-1+x)(1+6x)",
         ];
 
-        _.each(forms, function(form) {
+        _.each(forms, function (form) {
             expect(forms[0]).toHaveNorm(form);
         });
     });
@@ -48,7 +49,7 @@ describe("checking form", () => {
         expect("ab").toHaveStripNorm("ba");
         expect("(ab)c").toHaveStripNorm("(cb)a");
 
-        var forms = [
+        const forms = [
             "(6x+1)(x-1)",
             "(1+6x)(x-1)",
             "(6x+1)(-1+x)",
@@ -66,10 +67,10 @@ describe("checking form", () => {
             "(1-x)(-6x-1)",
             "(1-x)(-1-6x)",
             "-(6x+1)(1-x)",
-            "-(-6x-1)(x-1)"
+            "-(-6x-1)(x-1)",
         ];
 
-        _.each(forms, function(form) {
+        _.each(forms, function (form) {
             expect(forms[0]).toHaveStripNorm(form);
         });
     });