+++ packages/perseus-editor/src/iframe-content-renderer.jsx
@@ -12,10 +12,7 @@
  * its requestIframeData function in the parent, which triggers the parent to
  * send the current data.
  */
-
 import {Log} from "@khanacademy/perseus";
-import createReactClass from "create-react-class";
-import PropTypes from "prop-types";
 import * as React from "react";
 
 let nextIframeID = 0;
@@ -52,25 +49,32 @@ window.addEventListener("message", (event) => {
     }
 });
 
-const IframeContentRenderer: $FlowFixMe = createReactClass({
-    displayName: "IframeContentRenderer",
+type Props = {|
+    // The HTML content to render to the iframe
+    // content?: string,
 
-    propTypes: {
-        // The HTML content to render to the iframe
-        content: PropTypes.string.isRequired,
+    // The URL that the iframe should load
+    url: string,
 
-        // The data-* suffix for passing information to the iframe's JS
-        datasetKey: PropTypes.any,
+    // The data-* suffix for passing information to the iframe's JS
+    datasetKey: string,
 
-        // The value of the data-* attribute
-        datasetValue: PropTypes.any,
+    // The value of the data-* attribute
+    datasetValue: $FlowFixMe,
 
-        // Whether to make the iframe's height match its content's height,
-        // used to prevent scrolling inside the iframe.
-        seamless: PropTypes.bool,
-    },
+    // Whether to make the iframe's height match its content's height,
+    // used to prevent scrolling inside the iframe.
+    seamless: boolean,
+|};
 
-    componentDidMount: function () {
+class IframeContentRenderer extends React.Component<Props> {
+    _frame: ?HTMLIFrameElement;
+    _isMounted: boolean;
+    _lastData: $FlowFixMe;
+    _lastHeight: number;
+    iframeID: number;
+
+    componentDidMount() {
         // TODO(scottgrant): This is a hack to remove the deprecated call to
         // this.isMounted() but is still considered an anti-pattern.
         this._isMounted = true;
@@ -90,17 +94,16 @@ const IframeContentRenderer: $FlowFixMe = createReactClass({
                 this.refs.container.style.height = height + "px";
             }
         };
-    },
+    }
 
-    shouldComponentUpdate: function (nextProps) {
+    shouldComponentUpdate(nextProps: Props): boolean {
         return (
-            nextProps.content !== this.props.content ||
             nextProps.datasetValue !== this.props.datasetValue ||
             nextProps.seamless !== this.props.seamless
         );
-    },
+    }
 
-    componentDidUpdate: function (prevProps) {
+    componentDidUpdate(prevProps: Props) {
         if (!this.props.seamless) {
             // eslint-disable-next-line react/no-string-refs
             this.refs.container.style.height = "100%";
@@ -109,41 +112,38 @@ const IframeContentRenderer: $FlowFixMe = createReactClass({
             this.refs.container.style.height = this._lastHeight + "px";
         }
 
-        if (
-            prevProps.content !== this.props.content ||
-            prevProps.datasetValue !== this.props.datasetValue
-        ) {
+        if (prevProps.datasetValue !== this.props.datasetValue) {
             // Not just a change in seamless
             this._prepareFrame();
         }
-    },
+    }
 
-    componentWillUnmount: function () {
+    componentWillUnmount() {
         requestIframeData[this.iframeID] = null;
         updateIframeHeight[this.iframeID] = null;
 
         this._isMounted = false;
-    },
+    }
 
-    _prepareFrame: function () {
+    _prepareFrame() {
         // Don't initialize the iframe until the page has loaded
         if (this._frame) {
             // eslint-disable-next-line react/no-string-refs
             this.refs.container.removeChild(this._frame);
         }
 
-        this._frame = document.createElement("iframe");
-        this._frame.style.width = "100%";
-        this._frame.style.height = "100%";
+        const frame = document.createElement("iframe");
+        frame.style.width = "100%";
+        frame.style.height = "100%";
+        frame.src = this.props.url;
 
         if (this.props.datasetKey) {
             // If the user has specified a data-* attribute to place on the
             // iframe, we set it here. Right now, this is used to
             // communicate if the iframe should be enabling touch emulation.
-            this._frame.dataset[this.props.datasetKey] =
-                this.props.datasetValue;
+            frame.dataset[this.props.datasetKey] = this.props.datasetValue;
         }
-        this._frame.dataset.id = this.iframeID;
+        frame.dataset.id = String(this.iframeID);
 
         if (this.props.seamless) {
             // The seamless prop is the same as the "nochrome" prop that
@@ -152,50 +152,32 @@ const IframeContentRenderer: $FlowFixMe = createReactClass({
             // for lint indicators in the right margin. We use the dataset
             // as above to pass this information on to the perseus-frame
             // component inside the iframe
-            this._frame.dataset.lintGutter = "true";
+            frame.dataset.lintGutter = "true";
         }
 
-        // To make sure the value of location.href inside the iframe is the
-        // same as the location of the parent, we wait for the iframe to
-        // load before writing contents. Without the wait, the location
-        // inside the iframe becomes "about:blank", which causes problems
-        // with loading $LAB.
-        this._frame.onload = () => {
-            this._frame.onload = null;
-
-            // To prevent an issue with the contents of the iframe not being
-            // loaded properly, where the javascript inside the iframe is
-            // not executed, we push the content window write to the end
-            // of the event queue.
-            // TODO(jeff, CP-3128): Use Wonder Blocks Timing API
-            // eslint-disable-next-line no-restricted-syntax
-            setTimeout(() => {
-                this._frame.contentWindow.document.open();
-                this._frame.contentWindow.document.write(this.props.content);
-                this._frame.contentWindow.document.close();
-            });
-        };
-
         // eslint-disable-next-line react/no-string-refs
-        this.refs.container.appendChild(this._frame);
-    },
+        this.refs.container.appendChild(frame);
 
-    sendNewData: function (data) {
-        if (this._isMounted && data && this._frame.contentWindow) {
+        this._frame = frame;
+    }
+
+    sendNewData(data: $FlowFixMe) {
+        const frame = this._frame;
+        if (this._isMounted && data && frame?.contentWindow) {
             this._lastData = data;
 
             // We can't use JSON.stringify/parse for this because the apiOptions
             // includes the functions GroupMetadataEditor, groupAnnotator,
             // onFocusChange, and onInputError.
             window.iframeDataStore[this.iframeID] = data;
-            this._frame.contentWindow.postMessage(this.iframeID, "*");
+            frame.contentWindow.postMessage(this.iframeID, "*");
         }
-    },
+    }
 
-    render: function () {
+    render(): React.Node {
         // eslint-disable-next-line react/no-string-refs
         return <div ref="container" style={{width: "100%", height: "100%"}} />;
-    },
-});
+    }
+}
 
 export default IframeContentRenderer;