+++ packages/perseus-linter/src/types.js
@@ -1,0 +1,10 @@
+// @flow
+
+export type LinterContextProps = {
+    contentType: string,
+    highlightLint: boolean,
+    paths: $ReadOnlyArray<string>,
+    stack: $ReadOnlyArray<string>,
+    // additional properties can be added to the context by widgets
+    ...
+};