+++ packages/perseus/src/widgets/explanation.jsx
@@ -1,10 +1,10 @@
 /* eslint-disable react/sort-comp */
 // @flow
+import {linterContextDefault} from "@khanacademy/perseus-linter";
 import {StyleSheet, css} from "aphrodite";
 import * as React from "react";
 import _ from "underscore";
 
-import {linterContextDefault} from "../gorgon/proptypes.js";
 import * as Changeable from "../mixins/changeable.jsx";
 import Renderer from "../renderer.jsx";
 import * as styleConstants from "../styles/constants.js";