+++ packages/perseus/src/widgets/sorter.jsx
@@ -1,10 +1,13 @@
 // @flow
+import {
+    linterContextProps,
+    linterContextDefault,
+} from "@khanacademy/perseus-linter";
 import PropTypes from "prop-types";
 import * as React from "react";
 import _ from "underscore";
 
 import Sortable from "../components/sortable.jsx";
-import {linterContextProps, linterContextDefault} from "../gorgon/proptypes.js";
 import {ApiOptions} from "../perseus-api.jsx";
 import Util from "../util.js";