+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @khanacademy/perseus-editor
 
+## 0.3.4
+
+### Patch Changes
+
+-   Updated dependencies [a4f10ace]
+    -   @khanacademy/perseus@0.2.8
+
+## 0.3.3
+
+### Patch Changes
+
+-   Updated dependencies [3921a385]
+    -   @khanacademy/perseus@0.2.7
+
+## 0.3.2
+
+### Patch Changes
+
+-   5d67bb03: Add missing deps to peer/dev deps in package.json files
+-   Updated dependencies [5d67bb03]
+    -   @khanacademy/perseus@0.2.6
+
+## 0.3.1
+
+### Patch Changes
+
+-   @khanacademy/perseus@0.2.5
+
+## 0.3.0
+
+### Minor Changes
+
+-   8315e911: Don't re-export Perseus from @khanacademy/perseus-editor
+
+### Patch Changes
+
+-   Updated dependencies [5d86a625]
+    -   @khanacademy/perseus@0.2.4
+
+## 0.2.0
+
+### Minor Changes
+
+-   309cd98d: Content editor components must now specify a previewURL instead of frameSource
+
+### Patch Changes
+
+-   Updated dependencies [73095170]
+    -   @khanacademy/perseus@0.2.3
+
 ## 0.1.3
 
 ### Patch Changes